GENAI/Code/TRANSFORMER_MODELS/VISION/1.ImageClassification/GoogleVit/googlevit-base/requirements.txt
torch
transformers
Pillow
gradio
